Initial commit - everything runs locally
This supports saving and retreiving the two data types, and performing a
run on a single user account, including pulling configured modules,
determining which entries to send, building the email, and sending it.

So far it has only been testing with a single single-entry module.
Further testing will be required. I'll also need to rig up the AWS
lambda email sender. And I should probably have a CLI script for
creating modules and entries. Oh, and unit tests.