Initial ruby code
This can bootstrap a new cloud formation stack (suffixed -ruby) and can
copy a basic "Hello, World" script and a vendor directory. Presumably my
own library directories will be included as well.

This is still using Grunt to manage packaging and interacting with the
stack.

And it improves the stack creation model because the model needs code to
present before it can create the lambda but the code needs the s3 bucket
to exist to be uploaded. So the create step now creates the bucket
,builds and uploads the code, and then updates the stack with the rest
of the resources.