Adding ActiveRecord support and EmailModules
This renames the "module" concept to "EmailModules" to because Ruby has
a standard library "Module" class and classes are open, so stating
`class Module` re-opens the existing "Module" class instead of creating
a new one. Also "EmailModule" makes it clear what type of "Module" we
are talking about.

I'm also updating the design to use ActiveRecord-style models instead of
a DataStore that generates *POJOs*. Since this is Ruby and not
Java(Script), I'm going to try to stick with the style that's used, or
at least used in Rails. It's a bit unusual and requires, as far as I can
see it, some class-level variables and methods, but I might try to poke
around at that and see if I can avoid it (because I'd rather avoid it).