Slight refactor to a mono-repo for an email-sender system
The overall system will include both the existing email-sender system
and the tools, but it will also include the upcoming web app; which will
be written in go.

The reason to take the mono-repo approach is to better handle the fact
that the CloudFormation specification is defined here, and all code
resources eventually map to a CF resource.

Overall, this means each lambda can have its own language, and the
general rule is that each lambda will be defined in its own directory,
or in the case of `go`, top-level file.