Adding an API Gateway entry point for the web app
This is just a basic API entry that supports a single, optional,
parameter that is passed into the function. The function now returns a
json struct.

While this isn't the ideal end state -> I probably want a POST so I can
upload json; and while that's not true REST, I can't bring myself to
care. But we'll see.