Removes go webapp endpoint
After second thought, I'm going to keep everything in ts *and* keep
all actions local for now, or, if not local, use other mechanisms for
interacting with data.

If I ever want to make this go big, I'll probably just use elixir and
phoenix anyway.

The go-web code can be found on the go-web branch.