Adds initial dataPull lambda code
This lambda will be used to register and pull data for users that will
be included in modules. I'm running this in Python because something
just doesn't feel right about JS/TS, and now that both the Python2 thing
is basically wrapped up and that I'll be using Python at work, it is a
good time to return to it as an option.