Introduce the concept of module types
Each module will now have a `type` field that defines how it will be
processed. The default `type` is what exists today and will be called
`entries` (based on the fact that it reads from the `entries` field in
the Module record).

This also adds the ModuleProcessor process that will determine how a
particular module will be processed based on its type. It is currently
wired up tying the `entries` type to the current processing process.