[#12] Run against all users
This provides support for running the emailSender lambda against all
users, generating separate emails for each.

This uses a simple mechanism that pulls all users and kicks off
sepearate processes for each of them. This won't scale to large numbers
of users, but should be fine for the number we have.